refactor(tabbar): move TabBar related code from SceneDelegate to a separate HSTabBarViewController class
