+++ HubScout/Custom Views/TabbarControllers/HSTabBarViewController.swift
@@ -1,0 +1,34 @@
+//
+//  HSTabbarViewController.swift
+//  HubScout
+//
+//  Created by o9tech on 07/08/2024.
+//
+
+import UIKit
+
+class HSTabBarViewController: UITabBarController {
+
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        UITabBar.appearance().tintColor = .systemGreen
+        viewControllers                 = [createSearchNC(), createFavoritesNC()]
+    }
+
+
+    func createSearchNC() -> UINavigationController {
+        let searchVC = SearchVC()
+        searchVC.title = "Search"
+        searchVC.tabBarItem = UITabBarItem(tabBarSystemItem: .search, tag: 0)
+        return UINavigationController(rootViewController: searchVC)
+    }
+
+
+    func createFavoritesNC() -> UINavigationController {
+        let favoritesListVC = FavoritesListVC()
+        favoritesListVC.title = "Favorites"
+        favoritesListVC.tabBarItem = UITabBarItem(tabBarSystemItem: .favorites, tag: 1)
+        return UINavigationController(rootViewController: favoritesListVC)
+    }
+}
